fix verify user based on jwt
Fix access token and refresh token verification. The verification syntax differs between them. The refreshToken verification requires await since it runs in the controller. The verify() function has a callback that conflicts with try unless set as an async function. However, accessToken verification is in middleware, so its flow is correct.